requirements.txt
httpx==0.27.2
pandas==2.2.3
plotly==5.24.1
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
PyYAML==6.0.2
SQLAlchemy==2.0.36
streamlit==1.39.0
